+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # {{next-version}}
 
+* Guarantee (and document) the order of execution of commands created by merged video options of the `.playVideo()` method. Async first, then sync, then goto.
 
 ## v0.2.0